// LDC.Atlas.WebApp/Client/app/shared/enums/accounting-fields.enum.ts
export enum AccountingFields {
    DefaultBankAccountId = 'defaultBankAccountId',
    SalesLedgerControlClientDebtorsId = 'salesLedgerControlClientDebtorsId',
    PurchaseLedgerControlClientCreditorsId = 'purchaseLedgerControlClientCreditorsId',
    VATAccountInputs = 'vatAccountInputs',
    VATAccountOutputs = 'vatAccountOutputs',
    FXRevalaccountId = 'fxRevalaccountId',
    FXAccountGain = 'fxAccountGain',
    FXAccountLoss = 'fxAccountLoss',
    SuspenseAccountforWashoutSuspenseId = 'suspenseAccountforWashoutSuspenseId',
    RealisedPhysicalsPayableId = 'realisedPhysicalsPayableId',
    RealisedPhysicalsReceivableId = 'realisedPhysicalsReceivableId',
    CashReceivedCostTypeId = 'cashReceivedCostTypeId',
    CashPaidCostTypeId = 'cashPaidCostTypeId',
    PurchaseInvoiceCostTypeId = 'purchaseInvoiceCostTypeId',
    SalesInvoiceCostTypeId = 'salesInvoiceCostTypeId',
    WashoutInvoiceGainsCostTypeId = 'washoutInvoiceGainsCostTypeId',
    WashoutInvoiceLossCostTypeId = 'washoutInvoiceLossCostTypeId',
    FXRevalCostTypeId = 'fxRevalCostTypeId',
    PLClearanceYepAccountId = 'plClearanceYepAccountId',
    BalanceSheetClearanceYepAccountId = 'balanceSheetClearanceYepAccountId',
    BSReserveYepAccountId = 'bsReserveYepAccountId',
    YepCostTypeId = 'yepCostTypeId',
    DealId = 'dealId',
    SettlementId = 'settlementId',
    CancellationGainCostTypeId = 'cancellationGainCostTypeId',
    CancellationLossCostTypeId = 'cancellationLossCostTypeId',
}